requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
chromadb==0.5.5
google-generativeai==0.8.1
python-dotenv==1.0.1
openpyxl==3.1.5
lxml==5.3.0
